+++ src/components/HomePage.js
@@ -1,33 +1,23 @@
-import React, { useState } from 'react';
+import React from 'react';
 import foodBackground from '../images/berry-background.jpg';
-import Result from './Result';
 import Chart from './Chart';
-const axios = require('axios');
-const APP_ID = '7b632c32';
-const APP_KEY = '005154710d7c48250feb6e1dbd9bd7d6';
 
-function HomePage() {
-  const [searchState, setSearchState] = useState('');
-  const [foodData, setFoodData] = useState();
+function HomePage(props) {
+  console.log(props.foodData.calories);
 
-  const apiGetLink = `https://api.edamam.com/api/nutrition-data?app_id=${APP_ID}&app_key=${APP_KEY}&ingr=${searchState}`;
+  //Total Calories
+  const totalCalories = props.foodData.calories;
 
-  //axios get request for food API
-  const handleFoodData = () => {
-    axios
-      .get(apiGetLink)
-      .then((response) => {
-        console.log(response);
-        setFoodData((previousState) => response.data);
-      })
-      .catch((err) => console.error(err));
-  };
+  // Total Carbs
+  const totalCarbs = props.foodData.totalNutrients.CHOCDF;
+  const tCarbs = totalCarbs.quantity.toFixed(2);
 
-  // When food item is submited this function will run
-  const onSubmit = (event) => {
-    event.preventDefault();
-    handleFoodData();
-  };
+  //Total Protein
+  const totalProtein = props.foodData.totalNutrients.PROCNT;
+  const tProtein = totalProtein.quantity.toFixed(2);
+
+
+
 
   return (
     <div
@@ -46,7 +36,6 @@ function HomePage() {
               <button
                 className="bg-offwhite hover:bg-turquoise text-darkgray text-2xl font-bold py-2 px-4 w-full rounded focus:outline-none focus:shadow-outline"
                 type="submit"
-                onClick={onSubmit}
               >
                 Log Food <i className="fas fa-plus-circle"></i>
               </button>
@@ -59,8 +48,6 @@ function HomePage() {
               <ul className="bg-gray p-3">
                 <li>
                   <textarea
-                    onChange={(event) => setSearchState(event.target.value)}
-                    value={searchState}
                     cols="70"
                     rows="1"
                     className="border p-3 bg-offwhite"
@@ -76,12 +63,17 @@ function HomePage() {
               <div className="bg-tangerine">
                 <Chart />
               </div>
-
             </div>
           </form>
         </div>
       </section>
-      {foodData && <Result foodData={foodData} />}
+      <p>Calories: {totalCalories}</p>
+      <p>
+        {totalCarbs.label}: {tCarbs + totalCarbs.unit}
+      </p>
+      <p>
+        {totalProtein.label}: {tProtein + totalProtein.unit}
+      </p>
     </div>
   );
 }
